6.RAG-pipline/lambda/requirements.txt
langchain
langchain-community

aws-psycopg2
tiktoken
pypdf
